Factor out allowed package retrieval into a separate manager class
